Added fixes for MooBench table style
It adds two classes ``moobench-table-label`` and ``moobench-table-value``.
The added classes will look for the appropriate style in style.css.  The
original WordPress theme for kieker-monitoring.net is uploaded on the
new https://github.com/shinhyungyang/wp-theme-kieker repository, and the
fixes for the MooBench table style is committed as the
``fix-moobench-table-style`` branch on the repository.